refactor: move watcher tests to a separate file for better organization
